# contracts/ras-adapter/openapi.yaml
openapi: 3.0.3
info:
  title: RAS Adapter API
  description: |
    REST API для управления кластерами 1С через RAS (Remote Administration Server).

    ## API Versions

    ### v2 (Recommended)
    Action-based API с query параметрами. Все endpoints используют формат:
    - `GET /api/v2/<action>?param1=X&param2=Y`
    - `POST /api/v2/<action>?param1=X&param2=Y`

    ### v1 (Deprecated)
    RESTful API с path параметрами. **DEPRECATED since 2025-11-27**.
    Будет удалён 2026-03-01. Используйте v2.

    ## Операции

    - Управление кластерами
    - Управление информационными базами
    - Блокировка/разблокировка баз
    - Управление сессиями

    **КРИТИЧНО**: Используйте `cluster_id` (не `cluster`) для всех endpoints.
  version: 2.0.0
  contact:
    name: CommandCenter1C Team
    url: https://github.com/commandcenter1c/commandcenter
servers:
  - url: http://localhost:8088
    description: Local development server
  - url: http://ras-adapter:8088
    description: Docker environment
...
paths:
...
  /api/v2/list-sessions:
    get:
      summary: List sessions
      description: Returns list of active user sessions for specified infobase
      operationId: listSessionsV2
      tags:
        - sessions-v2
      parameters:
        - name: cluster_id
          in: query
          required: true
          description: Cluster UUID
          schema:
            type: string
            format: uuid
        - name: infobase_id
          in: query
          required: true
          description: Infobase UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/terminate-session:
    post:
      summary: Terminate single session
      description: |
        Terminates a specific user session by ID.
        **Idempotent**: Returns success even if session already terminated.
      operationId: terminateSessionV2
      tags:
        - sessions-v2
      parameters:
        - name: cluster_id
          in: query
          required: true
          description: Cluster UUID
          schema:
            type: string
            format: uuid
        - name: session_id
          in: query
          required: true
          description: Session UUID to terminate
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session terminated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminateSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/terminate-sessions:
    post:
      summary: Terminate multiple sessions
      description: |
        Terminates all sessions for specified infobase.
        If `session_ids` provided in body - terminates only those sessions (NOT YET IMPLEMENTED).
        If body is empty - terminates ALL sessions for the infobase.
      operationId: terminateSessionsV2
      tags:
        - sessions-v2
      parameters:
        - name: cluster_id
          in: query
          required: true
          description: Cluster UUID
          schema:
            type: string
            format: uuid
        - name: infobase_id
          in: query
          required: true
          description: Infobase UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminateSessionsRequestV2'
      responses:
        '200':
          description: Sessions terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminateSessionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: Selective termination not implemented (when session_ids provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # API V1 - DEPRECATED (keeping for backward compatibility)
  # ============================================================================
...
  /api/v1/sessions:
    get:
      summary: "[DEPRECATED] List sessions"
      deprecated: true
      operationId: getSessions
      tags:
        - sessions-v1
      parameters:
        - name: cluster_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: infobase_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsResponse'
  /api/v1/sessions/terminate:
    post:
      summary: "[DEPRECATED] Terminate sessions"
      deprecated: true
      operationId: terminateSessions
      tags:
        - sessions-v1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminateSessionsRequest'
      responses:
        '200':
          description: Sessions terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminateSessionsResponse'
components:
  schemas:
...
    Session:
      type: object
      required:
        - uuid
        - user_name
        - application
      properties:
        uuid:
          type: string
          format: uuid
        session_id:
          type: string
          description: Legacy alias for UUID
        user_name:
          type: string
        application:
          type: string
        started_at:
          type: string
    SessionsResponse:
      type: object
      required:
        - sessions
        - count
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        count:
          type: integer
    TerminateSessionResponse:
      type: object
      required:
        - success
        - message
        - session_id
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Session terminated successfully"
        session_id:
          type: string
          format: uuid
    TerminateSessionsRequestV2:
      type: object
      properties:
        session_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Optional list of specific session IDs (NOT YET IMPLEMENTED)
    TerminateSessionsRequest:
      type: object
      required:
        - cluster_id
        - infobase_id
      properties:
        cluster_id:
          type: string
          format: uuid
        infobase_id:
          type: string
          format: uuid
        session_ids:
          type: array
          items:
            type: string
          description: Currently ignored - terminates ALL sessions
    TerminateSessionsResponse:
      type: object
      required:
        - terminated_count
      properties:
        terminated_count:
          type: integer
        failed_count:
          type: integer
        failed_sessions:
          type: array
          items:
            type: string

    # ============================================================================
    # COMMON SCHEMAS
    # ============================================================================
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'